Move SetTrueOrFalseOrNone to examples
