+++ examples/text-generation/run_generation.py
@@ -29,9 +29,16 @@ from pathlib import Path
 
 import torch
 from transformers import BatchEncoding
-from utils import adjust_batch, count_hpu_graphs, finalize_quantization, initialize_model, save_model
+from utils import (
+    SetTrueOrFalseOrNone,
+    adjust_batch,
+    count_hpu_graphs,
+    finalize_quantization,
+    initialize_model,
+    save_model,
+)
 
-from optimum.habana.utils import get_hpu_memory_stats, SetTrueOrFalseOrNone
+from optimum.habana.utils import get_hpu_memory_stats
 
 
 logging.basicConfig(
@@ -385,7 +392,9 @@ def setup_parser(parser):
         args.limit_hpu_graphs = False
 
     if args.use_flash_attention and args.flash_attention_fast_softmax is None:
-        logger.warning("`--flash_attention_fast_softmax` was not set; defaulting to True due to `--use_flash_attention` being enabled.")
+        logger.warning(
+            "`--flash_attention_fast_softmax` was not set; defaulting to True due to `--use_flash_attention` being enabled."
+        )
         args.flash_attention_fast_softmax = True
     else:
         args.flash_attention_fast_softmax = False
